--- a/include/ada/url_aggregator-inl.h
+++ b/include/ada/url_aggregator-inl.h
@@ -87,6 +87,26 @@ inline void url_aggregator::update_unencoded_base_hash(std::string_view input) {
   ADA_ASSERT_TRUE(validate());
 }
 
+ada_really_inline uint32_t url_aggregator::replace_and_resize(uint32_t start, uint32_t end, std::string_view input) {
+  uint32_t current_length = end - start;
+  uint32_t input_size = uint32_t(input.size());
+  uint32_t new_difference = input_size - current_length;
+
+  if (current_length == 0) {
+    buffer.insert(start, input);
+  } else if (input_size == current_length) {
+    buffer.replace(start, input_size, input);
+  } else if (input_size < current_length) {
+    buffer.erase(start, current_length - input_size);
+    buffer.replace(start, input_size, input);
+  } else {
+    buffer.replace(start, current_length, input.substr(0, current_length));
+    buffer.insert(start + current_length, input.substr(current_length));
+  }
+
+  return new_difference;
+}
+
 inline void url_aggregator::update_base_hostname(const std::string_view input) {
   ada_log("url_aggregator::update_base_hostname ", input, " [", input.size(), " bytes], buffer is '", buffer, "' [", buffer.size()," bytes]");
   ADA_ASSERT_TRUE(validate());
@@ -95,21 +115,13 @@ inline void url_aggregator::update_base_hostname(const std::string_view input) {
   // This next line is required for when parsing a URL like `foo://`
   add_authority_slashes_if_needed();
 
-  bool has_credential = components.protocol_end + 2 < components.host_start;
-  uint32_t current_length = components.host_end - components.host_start;
-  // next line could overflow but unsigned arithmetic has well-defined overflows.
-  uint32_t new_difference = uint32_t(input.size()) - current_length;
-  // The common case is current_length == 0.
-  buffer.erase(components.host_start, current_length);
+  bool has_credentials = components.protocol_end + 2 < components.host_start;
+  uint32_t new_difference = replace_and_resize(components.host_start, components.host_end, input);
 
-  uint32_t host_start = components.host_start;
-  // The common case is components.host_start == buffer.size().
-  if (has_credential) {
-    buffer.insert(host_start, "@");
-    host_start++;
+  if (has_credentials) {
+    buffer.insert(components.host_start, "@");
     new_difference++;
   }
-  buffer.insert(host_start, input);
   components.host_end += new_difference;
   components.pathname_start += new_difference;
   if (components.search_start != url_components::omitted) { components.search_start += new_difference; }